修改了边框闪烁问题

+++ Cyan-Stars/Assets/GamePlay/Scritps/OldScritps/GameManager.cs
@@ -22,7 +22,7 @@ public class GameManager : MonoBehaviour
     [Header("1.Combo数量")]
     public int combo = 0;//Combo数量
     [Header("2.分数")]
-    public float score = 0;//分数
+    public int score = 0;//分数
     [Header("3.评分")]
     public EvaluateType grade;//评分
     [Header("4.当前精准度")]
@@ -30,21 +30,21 @@ public class GameManager : MonoBehaviour
     [Header("5.各个音符的偏移")]
     public List<float> deviationList = new List<float>();//各个音符的偏移
     [Header("6.理论最高分")]
-    public float maxScore = 0;//理论最高分
+    public int maxScore = 0;//理论最高分
     [Header("7.各个评分数量")]
     public int excatNum = 0;
     public int greatNum = 0;
     public int rightNum = 0;
     public int badNum = 0;
     public int missNum = 0;
-    public void RefreshPlayingUI(int combo,float score,string grade)
+    public void RefreshPlayingUI(int combo,int score,string grade)
     {
         foreach(var item in playingUIList)
         {
             item.Refresh(combo,score,grade,-1);
         }
     }
-    public void RefreshData(int addCombo,float addScore,EvaluateType grade,float currentDeviation)
+    public void RefreshData(int addCombo,int addScore,EvaluateType grade,float currentDeviation)
     {
         if(addCombo < 0)
         {
